+++ sql/sql_base.cc
@@ -5657,7 +5657,7 @@ find_field_in_table(THD *thd, TABLE *table, const char *name, uint length,
         DBUG_EVALUATE_IF("test_completely_invisible", 0, 1))
       DBUG_RETURN((Field*)0);
 
-    *cached_field_index_ptr= field_ptr - table->field;
+    *cached_field_index_ptr= (uint)(field_ptr - table->field);
     field= *field_ptr;
   }
   else
